Added is_authenticated fields for UserLoginData and UserMembersData classes and changed main.py is_auth function Move tasks.db to db/ and refactored code in app.py Added logout button in templates/profile.html Added configuration in app.py and profile rendering Moved contents of forms.py and user.py in config.py and deleted forms.py and user.py
