+++ templates/profile.html
@@ -3,7 +3,10 @@
 {# Шаблон для отображения страницы "Профиль" #}
 
 {% block content %}
-    <h1>Команда {{ team }}</h1>
+    <h1>Команда {{ team_name }} <a href="/logout" type="button" class="btn btn-primary">Выйти из
+        аккаунта</a>
+    </h1>
+
 
     <h3>Место в рейтинге игры Пенальти: </h3>
     <h3>Место в рейтинге игры Домино: </h3>
